Make a tetris piece fall, finally!

+++ src/GameState/GridState.cpp
@@ -28,14 +28,14 @@ namespace Tetris::State
         return m_grid.at(y).at(x);
     }
 
-    size_t GridState::width() const
+    int GridState::width() const
     {
-        return m_grid.at(0).size();
+        return (int)m_grid.at(0).size();
     }
 
-    size_t GridState::height() const
+    int GridState::height() const
     {
-        return m_grid.size();
+        return (int)m_grid.size();
     }
 
     bool GridState::isInside(int x, int y) const
