+++ src/GameState/GridState.h
@@ -20,8 +20,8 @@ namespace Tetris::State
         [[nodiscard]] GridCellState &getCellAt(int x, int y);
         [[nodiscard]] bool isInside(int x, int y) const;
 
-        [[nodiscard]] size_t width() const;
-        [[nodiscard]] size_t height() const;
+        [[nodiscard]] int width() const;
+        [[nodiscard]] int height() const;
 
     private:
         std::vector<std::vector<GridCellState>> m_grid;
